Remove ScoreHandler
StereoKit Api only allows reading and writing files through
the FilePicker, which I believe is not a feature that suits
the game, to read/write the user needs to select the file
where the score will be written.
Even performing the action through onCancel and forcing pop-up
to close before the user does anything does not work as expected.