Add ScoreTracker
ScoreTracker is an abstract class that keeps track of the scores
when we invoke the initialise method we flag the ScoreTracker
so it listens for a new score as input.
When the game ends the score is provided and that flag is set to
false, so the Score won't be updated until the initialise method
is called again and resets such flag.